fix: working csr gpu sort implementation

+++ deliverable-1/project/csr-gpu-sort/include/config.h
@@ -37,7 +37,7 @@
 // #define DUMP_OUTPUT
 
 /*! \brief matrices and vectors data type */
-typedef float dtype_t;
+typedef double dtype_t;
 
 /*! \brief integer data types */
 typedef uint64_t dsize_t;
